refactor: give the `.running` rule its own module
`tab-snapshot.ts` joins a persisted tab snapshot against live sessions to
produce ROWS. `hasLiveEngineTab` is the single boolean built on the same
inputs — the one `get-task`/`collect` publish and that unattended cleanup
loops act on — and it had been riding along in that file.

Moving it out is also what keeps the file under the size cap after the
previous commit widened it: shrinking the explanation of a rule that decides
whether a loop deletes live work would have been the wrong trade, and
splitting `tab-snapshot.ts` along its read/write seam has to wait for the
`verbs-*` work in flight.

  before: wc -l packages/kobe/src/cli/api/tab-snapshot.ts -> 504 (over the 500 cap)
  after:  459 + 56 (task-running.ts)
