+++ packages/kobe/src/cli/api/runtime.ts
@@ -31,13 +31,13 @@ import { restoredTabLaunch } from "./tab-respawn.ts"
 import {
   type TaskSessionRow,
   closeTabsSnapshot,
-  hasLiveEngineTab,
   joinTaskTabs,
   markCliTabSession,
   mintCliTab,
   publishCliTabSnapshot,
   readTabsSnapshot,
 } from "./tab-snapshot.ts"
+import { hasLiveEngineTab } from "./task-running.ts"
 import {
   ApiError,
   type ApiRuntime,
